pyproject.toml
[tool.poetry]
name = "quantum-siren"
version = "0.1.0"
description = ""
authors = ["lc3267 <melvin.strobl@kit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pennylane = "^0.31.0"
pillow = "^10.0.0"
scikit-image = "^0.21.0"
ipykernel = "^6.25.0"
matplotlib = "^3.7.2"
torchvision = [
    {version = "^0.15.2+cpu", source = "torch_cpu", markers = "extra == 'cpu'"},
    {version = "^0.15.2", source = "torch_gpu", markers = "extra == 'gpu'"}
]
kedro = "^0.18.12"
torchsummary = "^1.5.1"
torchmetrics = "^1.0.2"

[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"

[[tool.poetry.source]]
name = "torch_gpu"
url = "https://download.pytorch.org/whl/cu118"

[tool.poetry.extras]
cpu=["torchvision"]
gpu=["torchvision"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.kedro]
package_name = "quantum_siren"
project_name = "quantum_siren"
kedro_init_version = "0.18.12"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/quantum_siren -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]
